# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for the B2B medical supplies marketplace
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders/checkout:
    post:
      operationId: checkoutCart
      tags: [orders]
      summary: Convert the current cart into an order
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithItems"
        "400":
          description: Cart is empty or invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order with its items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithItems"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (vendor/admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithItems"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /orders/{id}/reorder:
    post:
      operationId: reorderOrder
      tags: [orders]
      summary: Add all items from a past order back into the current cart (buy again)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Number of items re-added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReorderResult"
...
  /admin/reports/sales:
    get:
      operationId: getSalesReport
      tags: [reports]
      summary: Sales revenue/order trend report (admin)
      parameters:
        - name: days
          in: query
          required: false
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Sales report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesReport"
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Order:
      type: object
      required: [id, buyerUserId, buyerCompanyId, status, totalAmount, createdAt]
      properties:
        id:
          type: integer
        buyerUserId:
          type: string
        buyerCompanyId:
          type: [integer, "null"]
        status:
          type: string
          enum: [pending, confirmed, shipped, completed, cancelled]
        totalAmount:
          type: number
        createdAt:
          type: string
          format: date-time
    OrderItemDetail:
      type: object
      required: [id, orderId, offerId, productId, productName, vendorCompanyId, vendorName, quantity, unitPrice, subtotal]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        offerId:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        vendorCompanyId:
          type: integer
        vendorName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        subtotal:
          type: number
    OrderWithItems:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/OrderItemDetail"
...
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, confirmed, shipped, completed, cancelled]
...
    ReorderResult:
      type: object
      required: [itemsAdded]
      properties:
        itemsAdded:
          type: integer
...
    SalesReportPoint:
      type: object
      required: [date, revenue, orders]
      properties:
        date:
          type: string
        revenue:
          type: number
        orders:
          type: integer
    SalesReport:
      type: object
      required: [points, totalRevenue, totalOrders]
      properties:
        points:
          type: array
          items:
            $ref: "#/components/schemas/SalesReportPoint"
        totalRevenue:
          type: number
        totalOrders:
          type: integer
...